Port DeepCopy rank 7 tests
