+++ core/perf_test/CMakeLists.txt
@@ -34,10 +34,6 @@ SET(SOURCES
   PerfTest_ViewCopy_b6.cpp
   PerfTest_ViewCopy_c6.cpp
   PerfTest_ViewCopy_d6.cpp
-  PerfTest_ViewCopy_a7.cpp
-  PerfTest_ViewCopy_b7.cpp
-  PerfTest_ViewCopy_c7.cpp
-  PerfTest_ViewCopy_d7.cpp
   PerfTest_ViewAllocate.cpp
   PerfTest_ViewFill_123.cpp
   PerfTest_ViewFill_45.cpp
@@ -198,6 +194,11 @@ ENDFUNCTION()
 SET(
   BENCHMARK_SOURCES
   BenchmarkMain.cpp
+  Benchmark_ViewCopy.cpp
+  PerfTest_ViewCopy_a7.cpp
+  PerfTest_ViewCopy_b7.cpp
+  PerfTest_ViewCopy_c7.cpp
+  PerfTest_ViewCopy_d7.cpp
   PerfTest_ViewCopy_a8.cpp
   PerfTest_ViewCopy_b8.cpp
   PerfTest_ViewCopy_c8.cpp
